+++ NEWS.md
@@ -75,4 +75,5 @@
 * Fixed a problem in the estimation algorithm that occurred when estimating a structural model with zero constraints in the W matrix. 
 * Increased the default maxit from 300 to 500 in fitGMVAR, the default maxlag from 10 to 12 in diagnostic_plot, and decreased the default ar_scale from 1 to 0.5 in GAfit.
 * Updated the documentation of fitGMVAR: more instructions on what to do when the algorithm fails to create an initial population were added.
-
+* By default there is now randomly varying "ar_scale" in the genetic algorithm
+* There is now random preliminary smart mutations in the genetic algorithm